refactor(app2): restructure transfer service files
