+++ app2/src/lib/services/transfer/transactions.ts
@@ -1,0 +1,34 @@
+import { Effect } from "effect"
+import type {
+  Hash,
+  SendTransactionErrorType,
+  SendTransactionParameters,
+  WaitForTransactionReceiptErrorType
+} from "viem"
+import { SendTransactionError, WaitForTransactionReceiptError } from "./errors"
+import { getPublicClient, getWalletClient } from "./clients"
+
+export const submitTransfer = (transactionArgs: SendTransactionParameters) =>
+  Effect.gen(function* () {
+    const walletClient = yield* getWalletClient
+
+    const hash = yield* Effect.tryPromise({
+      try: () => walletClient.sendTransaction(transactionArgs),
+      catch: err => new SendTransactionError({ cause: err as SendTransactionErrorType })
+    })
+
+    return hash
+  })
+
+export const waitForReceipt = (hash: Hash) =>
+  Effect.gen(function* () {
+    const publicClient = yield* getPublicClient
+
+    const receipt = yield* Effect.tryPromise({
+      try: () => publicClient.waitForTransactionReceipt({ hash }),
+      catch: err =>
+        new WaitForTransactionReceiptError({ cause: err as WaitForTransactionReceiptErrorType })
+    })
+
+    return receipt
+  })
